Taking action arguments from workflow
